fix: [SC-26223] 🐛 Fix build error
